+++ MCBO/mcbo/optimizers/manual/casmopolitan.py
@@ -96,7 +96,7 @@ class Casmopolitan(BoBase):
                  tr_succ_tol: Optional[int] = None,
                  tr_fail_tol: Optional[int] = None,
                  tr_verbose: bool = False,
-                 dtype: torch.dtype = torch.float32,
+                 dtype: torch.dtype = torch.float64,
                  device: torch.device = torch.device('cpu'),
                  ):
